+++ demo/admin/src/config.ts
@@ -1,8 +1,27 @@
+import cometConfig from "../comet-config.json";
 import environment from "./environment";
 
-const config: { [key in typeof environment[number]]: string } = environment.reduce((ret, value) => {
-    ret[value] = (window as unknown as Record<string, string>)[`EXTERNAL__${value}__`];
-    return ret;
-}, {} as Record<typeof environment[number], string>);
+const environmentVariables = {} as Record<typeof environment[number], string>;
+for (const variableName of environment) {
+    const externalVariableName = `EXTERNAL__${variableName}__`;
 
-export default config;
+    if (externalVariableName in window) {
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        environmentVariables[variableName] = (window as any)[externalVariableName];
+    } else {
+        // eslint-disable-next-line no-console
+        console.warn(`External variable ${externalVariableName} not set"`);
+    }
+}
+
+export const config = {
+    ...cometConfig,
+    apiUrl: environmentVariables.API_URL,
+    adminUrl: environmentVariables.ADMIN_URL,
+    sitesConfig: JSON.parse(environmentVariables.SITES_CONFIG),
+    buildDate: environmentVariables.BUILD_DATE,
+    buildNumber: environmentVariables.BUILD_NUMBER,
+    commitSha: environmentVariables.COMMIT_SHA,
+};
+
+export type Config = typeof config;