+++ packages/api/cms-api/generate-schema.ts
@@ -6,6 +6,7 @@ import { printSchema } from "graphql";
 
 import {
     BuildsResolver,
+    createAuthResolver,
     createPageTreeResolver,
     createRedirectsResolver,
     CurrentUserInterface,
@@ -18,7 +19,6 @@ import {
     PageTreeNodeBase,
     PageTreeNodeCategory,
 } from "./src";
-import { createAuthResolver } from "./src/auth/auth.resolver";
 import { BuildTemplatesResolver } from "./src/builds/build-templates.resolver";
 import { CronJobsResolver } from "./src/cron-jobs/cron-jobs.resolver";
 import { DamItemsResolver } from "./src/dam/files/dam-items.resolver";
@@ -83,7 +83,7 @@ async function generateSchema(): Promise<void> {
         PageTreeNode,
         Documents: [Page],
     }); // no scope
-    const AuthResolver = createAuthResolver(CurrentUser);
+    const AuthResolver = createAuthResolver({ currentUser: CurrentUser });
 
     const schema = await gqlSchemaFactory.create([
         BuildsResolver,